modules/location-producer/requirements.txt
kafka-python
keyboard
grpcio-tools